Split gpu_renderer, decode_worker, worker_api into directory modules
Phase 4: Split the three 1300+ line nexrad monolith files into focused
sub-modules within directory modules.

gpu_renderer/ (was 1307 lines):
- mod.rs: RadarGpuRenderer struct, new(), draw()
- textures.rs: texture creation/upload, color table management
- shaders.rs: shader sources and build functions (absorbs shader_common.rs)
- inspect.rs: value_at_polar, detect_storm_cells

decode_worker/ (was 1336 lines):
- mod.rs: DecodeWorker struct, lifecycle, re-exports
- types.rs: IngestResult, DecodeResult, VolumeData, WorkerOutcome, etc.
- send.rs: ingest/render/render_live/render_volume send methods
- receive.rs: onmessage callback, result deserialization

worker_api/ (was 1342 lines):
- mod.rs: WORKER_IDB thread-local, shared types
- ingest.rs: worker_ingest, worker_ingest_chunk WASM exports
- render.rs: worker_render, worker_render_volume exports
- render_live.rs: worker_render_live export
