+++ src/nexrad/decode_worker/mod.rs
@@ -1,0 +1,255 @@
+//! Web Worker-based data operations.
+//!
+//! Offloads expensive NEXRAD operations (ingestion, rendering) from the main UI
+//! thread into a dedicated Web Worker. Communication uses `postMessage` with
+//! Transferable ArrayBuffers for zero-copy data transfer.
+
+mod receive;
+mod send;
+mod types;
+
+pub use types::{
+    ChunkIngestContext, ChunkIngestResult, DecodeResult, IngestContext, IngestResult,
+    RenderContext, VolumeData, VolumeRenderContext, VolumeSweepMeta, WorkerOutcome,
+};
+
+use std::cell::RefCell;
+use std::collections::HashMap;
+use std::rc::Rc;
+use wasm_bindgen::JsCast;
+use web_sys::{Worker, WorkerOptions, WorkerType};
+
+use types::RequestId;
+
+/// Manages a dedicated Web Worker for NEXRAD data operations.
+///
+/// Created once at app startup and kept alive for the entire session.
+/// Supports two command types:
+/// - `ingest`: Split, probe, and store archive records in IDB
+/// - `render`: Selectively decode + render a single elevation
+///
+/// Results are polled via `try_recv()` each frame.
+pub struct DecodeWorker {
+    worker: Worker,
+    next_id: u64,
+    ready: Rc<RefCell<bool>>,
+    pending_ingest: Rc<RefCell<HashMap<RequestId, IngestContext>>>,
+    pending_chunk_ingest: Rc<RefCell<HashMap<RequestId, ChunkIngestContext>>>,
+    pending_render: Rc<RefCell<HashMap<RequestId, RenderContext>>>,
+    pending_render_live: Rc<RefCell<HashMap<RequestId, RenderContext>>>,
+    pending_volume: Rc<RefCell<HashMap<RequestId, VolumeRenderContext>>>,
+    results: Rc<RefCell<Vec<WorkerOutcome>>>,
+    /// Requests queued before the worker was ready.
+    queue: Vec<QueuedRequest>,
+}
+
+/// A request queued before the worker was ready.
+enum QueuedRequest {
+    Ingest(RequestId, Vec<u8>, String, i64, String),
+    IngestChunk(RequestId, Vec<u8>, String, i64, u32, bool, bool, String),
+    Render(RequestId, String, u8, String),
+    RenderLive(RequestId, u8, String),
+    RenderVolume(RequestId, String, String, Vec<u8>),
+}
+
+impl DecodeWorker {
+    /// Create a new decode worker.
+    ///
+    /// Discovers the WASM/JS URLs from the current page's `<link>` tags
+    /// (generated by Trunk), creates a module worker, and sends the init message.
+    /// The worker will post a "ready" message once WASM is initialized.
+    pub fn new(ctx: eframe::egui::Context) -> Result<Self, String> {
+        let js_url =
+            discover_js_url().ok_or_else(|| "Could not find JS module URL in DOM".to_string())?;
+        let wasm_url =
+            discover_wasm_url().ok_or_else(|| "Could not find WASM URL in DOM".to_string())?;
+
+        log::info!(
+            "Creating decode worker with JS={}, WASM={}",
+            js_url,
+            wasm_url
+        );
+
+        // Create an ES module worker
+        let mut opts = WorkerOptions::new();
+        #[allow(deprecated)]
+        opts.type_(WorkerType::Module);
+
+        let worker = Worker::new_with_options("worker.js", &opts)
+            .map_err(|e| format!("Failed to create Worker: {:?}", e))?;
+
+        let ready = Rc::new(RefCell::new(false));
+        let pending_ingest: Rc<RefCell<HashMap<RequestId, IngestContext>>> =
+            Rc::new(RefCell::new(HashMap::new()));
+        let pending_chunk_ingest: Rc<RefCell<HashMap<RequestId, ChunkIngestContext>>> =
+            Rc::new(RefCell::new(HashMap::new()));
+        let pending_render: Rc<RefCell<HashMap<RequestId, RenderContext>>> =
+            Rc::new(RefCell::new(HashMap::new()));
+        let pending_render_live: Rc<RefCell<HashMap<RequestId, RenderContext>>> =
+            Rc::new(RefCell::new(HashMap::new()));
+        let pending_volume: Rc<RefCell<HashMap<RequestId, VolumeRenderContext>>> =
+            Rc::new(RefCell::new(HashMap::new()));
+        let results: Rc<RefCell<Vec<WorkerOutcome>>> = Rc::new(RefCell::new(Vec::new()));
+
+        // Set up the onmessage handler
+        receive::setup_onmessage(
+            &worker,
+            &ctx,
+            &ready,
+            &pending_ingest,
+            &pending_chunk_ingest,
+            &pending_render,
+            &pending_render_live,
+            &pending_volume,
+            &results,
+        );
+
+        // Set up onerror handler
+        {
+            let onerror = wasm_bindgen::closure::Closure::<dyn Fn(web_sys::ErrorEvent)>::new(
+                move |event: web_sys::ErrorEvent| {
+                    log::error!(
+                        "Decode worker error: {} ({}:{})",
+                        event.message(),
+                        event.filename(),
+                        event.lineno()
+                    );
+                },
+            );
+
+            worker.set_onerror(Some(onerror.as_ref().unchecked_ref()));
+            onerror.forget();
+        }
+
+        // Send init message with the WASM/JS URLs
+        let init_msg = js_sys::Object::new();
+        js_sys::Reflect::set(&init_msg, &"type".into(), &"init".into()).ok();
+        js_sys::Reflect::set(&init_msg, &"jsUrl".into(), &js_url.into()).ok();
+        js_sys::Reflect::set(&init_msg, &"wasmUrl".into(), &wasm_url.into()).ok();
+
+        worker
+            .post_message(&init_msg)
+            .map_err(|e| format!("Failed to send init message: {:?}", e))?;
+
+        Ok(Self {
+            worker,
+            next_id: 1,
+            ready,
+            pending_ingest,
+            pending_chunk_ingest,
+            pending_render,
+            pending_render_live,
+            pending_volume,
+            results,
+            queue: Vec::new(),
+        })
+    }
+
+    fn next_request_id(&mut self) -> u64 {
+        let id = self.next_id;
+        self.next_id += 1;
+        id
+    }
+
+    /// Flush any queued requests if the worker has become ready.
+    pub fn flush_queue(&mut self) {
+        if *self.ready.borrow() && !self.queue.is_empty() {
+            let queued: Vec<_> = self.queue.drain(..).collect();
+            log::info!("Flushing {} queued worker requests", queued.len());
+            for request in queued {
+                match request {
+                    QueuedRequest::Ingest(id, data, site_id, ts, file_name) => {
+                        send::send_ingest_request(
+                            &self.worker,
+                            id,
+                            &data,
+                            &site_id,
+                            ts,
+                            &file_name,
+                        );
+                    }
+                    QueuedRequest::IngestChunk(
+                        id,
+                        data,
+                        site_id,
+                        ts,
+                        chunk_index,
+                        is_start,
+                        is_end,
+                        file_name,
+                    ) => {
+                        send::send_ingest_chunk_request(
+                            &self.worker,
+                            id,
+                            &data,
+                            &site_id,
+                            ts,
+                            chunk_index,
+                            is_start,
+                            is_end,
+                            &file_name,
+                        );
+                    }
+                    QueuedRequest::Render(id, scan_key, elev, product) => {
+                        send::send_render_request(&self.worker, id, &scan_key, elev, &product);
+                    }
+                    QueuedRequest::RenderLive(id, elev, product) => {
+                        send::send_render_live_request(&self.worker, id, elev, &product);
+                    }
+                    QueuedRequest::RenderVolume(id, scan_key, product, elev_nums) => {
+                        send::send_render_volume_request(
+                            &self.worker,
+                            id,
+                            &scan_key,
+                            &product,
+                            &elev_nums,
+                        );
+                    }
+                }
+            }
+        }
+    }
+
+    /// Poll for completed worker results. Call this each frame in the update loop.
+    pub fn try_recv(&mut self) -> Vec<WorkerOutcome> {
+        self.flush_queue();
+        self.results.borrow_mut().drain(..).collect()
+    }
+}
+
+// ---------------------------------------------------------------------------
+// DOM discovery helpers
+// ---------------------------------------------------------------------------
+
+/// Discover the Trunk-generated JS module URL from DOM `<link rel="modulepreload">` tags.
+fn discover_js_url() -> Option<String> {
+    let document = web_sys::window()?.document()?;
+    let links = document
+        .query_selector_all("link[rel='modulepreload']")
+        .ok()?;
+    for i in 0..links.length() {
+        if let Some(el) = links.get(i) {
+            let el: &web_sys::Element = el.unchecked_ref();
+            if let Some(href) = el.get_attribute("href") {
+                // Find the main app module (not snippet/helper modules)
+                if href.contains("nexrad-workbench") && href.ends_with(".js") {
+                    return Some(href);
+                }
+            }
+        }
+    }
+    None
+}
+
+/// Discover the Trunk-generated WASM URL from DOM `<link rel="preload">` tags.
+fn discover_wasm_url() -> Option<String> {
+    let document = web_sys::window()?.document()?;
+    let links = document
+        .query_selector_all("link[rel='preload'][type='application/wasm']")
+        .ok()?;
+    if let Some(el) = links.get(0) {
+        let el: &web_sys::Element = el.unchecked_ref();
+        return el.get_attribute("href");
+    }
+    None
+}
