--- src/nexrad/gpu_renderer.rs
@@ -1,1307 +1,0 @@
-//! GPU-based radar renderer using WebGL2 shaders via glow.
-//!
-//! Renders polar radar data (azimuths x gates) directly on the GPU using a fragment
-//! shader that performs polar-to-Cartesian conversion and color lookup from a LUT texture.
-
-use super::color_table::{
-    build_reflectivity_lut, continuous_color_scale, product_from_str, product_value_range,
-};
-use crate::state::RenderProcessing;
-use glow::HasContext;
-use nexrad_render::Product;
-use std::sync::Arc;
-
-const VERTEX_SHADER: &str = r#"#version 300 es
-precision highp float;
-
-in vec2 a_position;
-out vec2 v_screen_pos;
-
-uniform vec2 u_viewport_size;
-
-void main() {
-    gl_Position = vec4(a_position, 0.0, 1.0);
-    // Convert NDC (-1..1) to pixel coordinates in egui convention (Y-down).
-    // WebGL NDC has Y-up, so flip Y so (0,0) = top-left to match the
-    // radar center coordinate passed from egui screen space.
-    vec2 uv = a_position * 0.5 + 0.5;
-    v_screen_pos = vec2(uv.x, 1.0 - uv.y) * u_viewport_size;
-}
-"#;
-
-fn build_flat_fragment_shader() -> String {
-    use super::shader_common::*;
-    format!(
-        r#"#version 300 es
-precision highp float;
-
-in vec2 v_screen_pos;
-out vec4 fragColor;
-
-uniform vec2 u_radar_center;       // radar center in screen pixels
-uniform float u_radar_radius;      // max coverage radius in screen pixels
-{FRAGMENT_PREAMBLE}
-// Processing uniforms
-uniform int u_data_age_indicator; // 0 or 1: desaturate oldest data behind sweep line
-
-// Sweep animation (dual-texture compositing)
-uniform sampler2D u_prev_data_tex;    // previous scan gate values (R32F, texture unit 3)
-uniform sampler2D u_prev_azimuth_tex; // previous scan azimuths (R32F, texture unit 4)
-uniform int u_sweep_enabled;       // 0 or 1
-uniform float u_sweep_azimuth;     // current sweep line angle in degrees
-uniform float u_sweep_start;       // azimuth where the sweep began collecting
-uniform float u_prev_offset;       // previous scan moment offset
-uniform float u_prev_scale;        // previous scan moment scale
-// Previous sweep spatial params (may differ from current sweep)
-uniform float u_prev_gate_count;
-uniform float u_prev_azimuth_count;
-uniform float u_prev_first_gate_km;
-uniform float u_prev_gate_interval_km;
-uniform float u_prev_max_range_km;
-
-{SAMPLE_DATA_P}
-{IS_VALID}
-{FIND_NEAREST_AZ_P}
-{FIND_BRACKET_AZ_P}
-void main() {{
-    vec2 delta = v_screen_pos - u_radar_center;
-    float dist_px = length(delta);
-    float dist_km = (dist_px / u_radar_radius) * u_max_range_km;
-
-    float azimuth_rad = atan(delta.x, -delta.y);
-    float azimuth_deg = mod(degrees(azimuth_rad) + 360.0, 360.0);
-
-    // Sweep animation: determine whether to sample previous or current texture.
-    // Must be computed before range/gate checks because the two textures may
-    // have different spatial extents (e.g. 0.5° at 460 km vs 0.9° at 298 km).
-    bool use_prev = false;
-    if (u_sweep_enabled == 1) {{
-        float swept_arc = mod(u_sweep_azimuth - u_sweep_start, 360.0);
-        float pixel_from_start = mod(azimuth_deg - u_sweep_start, 360.0);
-        use_prev = (pixel_from_start >= swept_arc);
-    }}
-
-    // Age-based desaturation: oldest quadrant (270°-360° behind sweep) fades out
-    float desat_factor = 0.0;
-    if (u_sweep_enabled == 1 && u_data_age_indicator == 1) {{
-        float age = mod(u_sweep_azimuth - azimuth_deg + 360.0, 360.0) / 360.0;
-        desat_factor = clamp((age - 0.75) / 0.25, 0.0, 1.0) * 0.9;
-    }}
-
-    // --- Unified sweep pipeline: select spatial params and samplers based on use_prev ---
-    float s_gate_count    = use_prev ? u_prev_gate_count    : u_gate_count;
-    float s_azimuth_count = use_prev ? u_prev_azimuth_count : u_azimuth_count;
-    float s_first_gate_km = use_prev ? u_prev_first_gate_km : u_first_gate_km;
-    float s_gate_interval = use_prev ? u_prev_gate_interval_km : u_gate_interval_km;
-    float s_max_range_km  = use_prev ? u_prev_max_range_km  : u_max_range_km;
-    float s_offset        = use_prev ? u_prev_offset        : u_offset;
-    float s_scale         = use_prev ? u_prev_scale         : u_scale;
-
-    // Range check
-    if (dist_km < s_first_gate_km || dist_km >= s_max_range_km) {{
-        fragColor = vec4(0.0);
-        return;
-    }}
-
-    float gate_idx = (dist_km - s_first_gate_km) / s_gate_interval;
-    if (gate_idx < 0.0 || gate_idx >= s_gate_count) {{
-        fragColor = vec4(0.0);
-        return;
-    }}
-
-    // Select the active samplers for the chosen sweep (current or previous).
-    // GLSL ES 3.00 requires sampler access in uniform control flow, but
-    // use_prev is uniform across the draw call for a given pixel's branch —
-    // we use explicit if/else to satisfy the compiler.
-    float value;
-
-    if (u_interpolation == 1) {{
-        // Bilinear interpolation
-        float az_lo, az_hi, az_frac;
-        bool bracket_ok;
-        if (use_prev) {{
-            bracket_ok = find_bracket_az_p(azimuth_deg, s_azimuth_count, u_prev_azimuth_tex, az_lo, az_hi, az_frac);
-        }} else {{
-            bracket_ok = find_bracket_az_p(azimuth_deg, s_azimuth_count, u_azimuth_tex, az_lo, az_hi, az_frac);
-        }}
-        if (!bracket_ok) {{
-            fragColor = vec4(0.0);
-            return;
-        }}
-
-        float g_lo = floor(gate_idx);
-        float g_hi = min(g_lo + 1.0, s_gate_count - 1.0);
-        float g_frac = gate_idx - g_lo;
-
-        float v00, v10, v01, v11;
-        if (use_prev) {{
-            v00 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_lo, az_lo);
-            v10 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_hi, az_lo);
-            v01 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_lo, az_hi);
-            v11 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_hi, az_hi);
-        }} else {{
-            v00 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_lo, az_lo);
-            v10 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_hi, az_lo);
-            v01 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_lo, az_hi);
-            v11 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_hi, az_hi);
-        }}
-
-        // Weighted average skipping sentinel values
-        float sum = 0.0;
-        float wsum = 0.0;
-        float w00 = (1.0 - g_frac) * (1.0 - az_frac);
-        float w10 = g_frac * (1.0 - az_frac);
-        float w01 = (1.0 - g_frac) * az_frac;
-        float w11 = g_frac * az_frac;
-
-        if (is_valid(v00)) {{ sum += v00 * w00; wsum += w00; }}
-        if (is_valid(v10)) {{ sum += v10 * w10; wsum += w10; }}
-        if (is_valid(v01)) {{ sum += v01 * w01; wsum += w01; }}
-        if (is_valid(v11)) {{ sum += v11 * w11; wsum += w11; }}
-
-        if (wsum < 0.001) {{
-            fragColor = vec4(0.0);
-            return;
-        }}
-        value = sum / wsum;
-    }} else {{
-        // Nearest neighbor
-        float dummy_az;
-        float best_idx;
-        if (use_prev) {{
-            best_idx = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_prev_azimuth_tex, dummy_az);
-        }} else {{
-            best_idx = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_azimuth_tex, dummy_az);
-        }}
-        if (best_idx < 0.0) {{
-            fragColor = vec4(0.0);
-            return;
-        }}
-        if (use_prev) {{
-            value = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, floor(gate_idx), best_idx);
-        }} else {{
-            value = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, floor(gate_idx), best_idx);
-        }}
-    }}
-
-    if (!is_valid(value)) {{
-        fragColor = vec4(0.0);
-        return;
-    }}
-
-    // Despeckle filter
-    if (u_despeckle_enabled == 1) {{
-        float dummy_az2;
-        float center_az;
-        if (use_prev) {{
-            center_az = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_prev_azimuth_tex, dummy_az2);
-        }} else {{
-            center_az = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_azimuth_tex, dummy_az2);
-        }}
-        float center_g = floor(gate_idx);
-        if (center_az >= 0.0) {{
-            int valid_count = 0;
-            for (int dg = -1; dg <= 1; dg++) {{
-                for (int da = -1; da <= 1; da++) {{
-                    if (dg == 0 && da == 0) continue;
-                    float ng = center_g + float(dg);
-                    float na = mod(center_az + float(da), s_azimuth_count);
-                    if (use_prev) {{
-                        if (is_valid(sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, ng, na))) valid_count++;
-                    }} else {{
-                        if (is_valid(sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, ng, na))) valid_count++;
-                    }}
-                }}
-            }}
-            if (valid_count < u_despeckle_threshold) {{
-                fragColor = vec4(0.0);
-                return;
-            }}
-        }}
-    }}
-
-{RAW_TO_PHYSICAL}
-{COLOR_LOOKUP}
-    // Age-based desaturation for sweep animation
-    if (desat_factor > 0.0) {{
-        float lum = dot(color.rgb, vec3(0.299, 0.587, 0.114));
-        color.rgb = mix(color.rgb, vec3(lum), desat_factor);
-    }}
-
-    // Apply opacity and output premultiplied alpha (egui requirement)
-    float a = color.a * u_opacity;
-    fragColor = vec4(color.rgb * a, a);
-}}
-"#
-    )
-}
-
-/// Find the nearest azimuth index in an array of azimuth angles.
-///
-/// Returns `None` if the nearest azimuth is farther than 1.5x the expected spacing
-/// (gap detection). Used by CPU-side inspector lookups.
-fn find_nearest_azimuth_index(
-    azimuths: &[f32],
-    azimuth_count: usize,
-    target_deg: f32,
-) -> Option<usize> {
-    let mut best_idx = 0usize;
-    let mut best_dist = 360.0f32;
-    for (i, &az) in azimuths.iter().enumerate() {
-        let mut d = (target_deg - az).abs();
-        if d > 180.0 {
-            d = 360.0 - d;
-        }
-        if d < best_dist {
-            best_dist = d;
-            best_idx = i;
-        }
-    }
-
-    let az_spacing = 360.0 / azimuth_count as f32;
-    if best_dist > az_spacing * 1.5 {
-        return None;
-    }
-
-    Some(best_idx)
-}
-
-/// All uniform locations for the radar shader program.
-struct UniformLocations {
-    radar_center: glow::UniformLocation,
-    radar_radius: glow::UniformLocation,
-    gate_count: glow::UniformLocation,
-    azimuth_count: glow::UniformLocation,
-    first_gate_km: glow::UniformLocation,
-    gate_interval_km: glow::UniformLocation,
-    max_range_km: glow::UniformLocation,
-    value_min: glow::UniformLocation,
-    value_range: glow::UniformLocation,
-    viewport_size: glow::UniformLocation,
-    interpolation: glow::UniformLocation,
-    despeckle_enabled: glow::UniformLocation,
-    despeckle_threshold: glow::UniformLocation,
-    opacity: glow::UniformLocation,
-    data_age_indicator: glow::UniformLocation,
-    offset: glow::UniformLocation,
-    scale: glow::UniformLocation,
-    sweep_enabled: glow::UniformLocation,
-    sweep_azimuth: glow::UniformLocation,
-    sweep_start: glow::UniformLocation,
-    prev_offset: glow::UniformLocation,
-    prev_scale: glow::UniformLocation,
-    prev_gate_count: glow::UniformLocation,
-    prev_azimuth_count: glow::UniformLocation,
-    prev_first_gate_km: glow::UniformLocation,
-    prev_gate_interval_km: glow::UniformLocation,
-    prev_max_range_km: glow::UniformLocation,
-}
-
-/// Spatial metadata for a single sweep (current or previous).
-struct SweepState {
-    azimuth_count: u32,
-    gate_count: u32,
-    first_gate_km: f64,
-    gate_interval_km: f64,
-    max_range_km: f64,
-    data_offset: f32,
-    data_scale: f32,
-    sweep_id: Option<String>,
-}
-
-impl Default for SweepState {
-    fn default() -> Self {
-        Self {
-            azimuth_count: 0,
-            gate_count: 0,
-            first_gate_km: 0.0,
-            gate_interval_km: 0.0,
-            max_range_km: 0.0,
-            data_offset: 0.0,
-            data_scale: 1.0,
-            sweep_id: None,
-        }
-    }
-}
-
-/// CPU-side shadow copies of sweep data for inspector/detection lookups.
-#[derive(Default)]
-struct CpuShadowData {
-    azimuths: Vec<f32>,
-    gate_values: Vec<f32>,
-    radial_times: Vec<f64>,
-}
-
-/// GPU-based radar renderer using WebGL2 shaders.
-pub struct RadarGpuRenderer {
-    program: glow::Program,
-    vao: glow::VertexArray,
-    #[allow(dead_code)] // Retained to prevent GPU resource deallocation
-    vbo: glow::Buffer,
-
-    data_texture: glow::Texture,
-    lut_texture: glow::Texture,
-    azimuth_texture: glow::Texture,
-
-    // Previous scan textures (for sweep animation)
-    prev_data_texture: glow::Texture,
-    prev_azimuth_texture: glow::Texture,
-
-    uniforms: UniformLocations,
-    current: SweepState,
-    prev: SweepState,
-    cpu: CpuShadowData,
-    prev_cpu: CpuShadowData,
-
-    has_data: bool,
-    value_min: f32,
-    value_range: f32,
-}
-
-impl RadarGpuRenderer {
-    /// Create a new GPU renderer, compiling shaders and allocating GL resources.
-    ///
-    /// Returns `Err` if shader compilation, program linking, or GL resource
-    /// allocation fails, allowing the caller to fall back gracefully.
-    pub fn new(gl: &Arc<glow::Context>) -> Result<Self, String> {
-        unsafe {
-            let program = gl
-                .create_program()
-                .map_err(|e| format!("Cannot create program: {}", e))?;
-
-            let vert = gl
-                .create_shader(glow::VERTEX_SHADER)
-                .map_err(|e| format!("Cannot create vertex shader: {}", e))?;
-            gl.shader_source(vert, VERTEX_SHADER);
-            gl.compile_shader(vert);
-            if !gl.get_shader_compile_status(vert) {
-                let info = gl.get_shader_info_log(vert);
-                return Err(format!("Vertex shader compile error: {}", info));
-            }
-
-            let frag = gl
-                .create_shader(glow::FRAGMENT_SHADER)
-                .map_err(|e| format!("Cannot create fragment shader: {}", e))?;
-            let frag_src = build_flat_fragment_shader();
-            gl.shader_source(frag, &frag_src);
-            gl.compile_shader(frag);
-            if !gl.get_shader_compile_status(frag) {
-                let info = gl.get_shader_info_log(frag);
-                return Err(format!("Fragment shader compile error: {}", info));
-            }
-
-            gl.attach_shader(program, vert);
-            gl.attach_shader(program, frag);
-            gl.link_program(program);
-            if !gl.get_program_link_status(program) {
-                let info = gl.get_program_info_log(program);
-                return Err(format!("Shader program link error: {}", info));
-            }
-            gl.detach_shader(program, vert);
-            gl.detach_shader(program, frag);
-            gl.delete_shader(vert);
-            gl.delete_shader(frag);
-
-            // Fullscreen quad (two triangles)
-            let vertices: [f32; 12] = [
-                -1.0, -1.0, 1.0, -1.0, 1.0, 1.0, -1.0, -1.0, 1.0, 1.0, -1.0, 1.0,
-            ];
-
-            let vbo = gl
-                .create_buffer()
-                .map_err(|e| format!("Cannot create VBO: {}", e))?;
-            gl.bind_buffer(glow::ARRAY_BUFFER, Some(vbo));
-            gl.buffer_data_u8_slice(
-                glow::ARRAY_BUFFER,
-                bytemuck_cast_slice(&vertices),
-                glow::STATIC_DRAW,
-            );
-
-            let vao = gl
-                .create_vertex_array()
-                .map_err(|e| format!("Cannot create VAO: {}", e))?;
-            gl.bind_vertex_array(Some(vao));
-            let a_position = gl
-                .get_attrib_location(program, "a_position")
-                .ok_or("Missing a_position")?;
-            gl.enable_vertex_attrib_array(a_position);
-            gl.vertex_attrib_pointer_f32(a_position, 2, glow::FLOAT, false, 8, 0);
-            gl.bind_vertex_array(None);
-
-            // Create placeholder textures (1x1)
-            let data_texture = create_r32f_texture(gl, 1, 1, &[0.0]);
-            let azimuth_texture = create_r32f_texture(gl, 1, 1, &[0.0]);
-            let lut_texture = create_rgba8_texture(gl, 1, 1, &[0, 0, 0, 0]);
-
-            // Helper to look up a required uniform location
-            let uniform = |name: &str| -> Result<glow::UniformLocation, String> {
-                gl.get_uniform_location(program, name)
-                    .ok_or_else(|| format!("Missing uniform: {}", name))
-            };
-
-            // Bind texture units to sampler uniforms
-            gl.use_program(Some(program));
-
-            let u_data_tex = uniform("u_data_tex")?;
-            gl.uniform_1_i32(Some(&u_data_tex), 0);
-            let u_lut_tex = uniform("u_lut_tex")?;
-            gl.uniform_1_i32(Some(&u_lut_tex), 1);
-            let u_azimuth_tex = uniform("u_azimuth_tex")?;
-            gl.uniform_1_i32(Some(&u_azimuth_tex), 2);
-            let u_prev_data_tex = uniform("u_prev_data_tex")?;
-            gl.uniform_1_i32(Some(&u_prev_data_tex), 3);
-            let u_prev_azimuth_tex = uniform("u_prev_azimuth_tex")?;
-            gl.uniform_1_i32(Some(&u_prev_azimuth_tex), 4);
-
-            let uniforms = UniformLocations {
-                radar_center: uniform("u_radar_center")?,
-                radar_radius: uniform("u_radar_radius")?,
-                gate_count: uniform("u_gate_count")?,
-                azimuth_count: uniform("u_azimuth_count")?,
-                first_gate_km: uniform("u_first_gate_km")?,
-                gate_interval_km: uniform("u_gate_interval_km")?,
-                max_range_km: uniform("u_max_range_km")?,
-                value_min: uniform("u_value_min")?,
-                value_range: uniform("u_value_range")?,
-                viewport_size: uniform("u_viewport_size")?,
-                interpolation: uniform("u_interpolation")?,
-                despeckle_enabled: uniform("u_despeckle_enabled")?,
-                despeckle_threshold: uniform("u_despeckle_threshold")?,
-                opacity: uniform("u_opacity")?,
-                data_age_indicator: uniform("u_data_age_indicator")?,
-                offset: uniform("u_offset")?,
-                scale: uniform("u_scale")?,
-                sweep_enabled: uniform("u_sweep_enabled")?,
-                sweep_azimuth: uniform("u_sweep_azimuth")?,
-                sweep_start: uniform("u_sweep_start")?,
-                prev_offset: uniform("u_prev_offset")?,
-                prev_scale: uniform("u_prev_scale")?,
-                prev_gate_count: uniform("u_prev_gate_count")?,
-                prev_azimuth_count: uniform("u_prev_azimuth_count")?,
-                prev_first_gate_km: uniform("u_prev_first_gate_km")?,
-                prev_gate_interval_km: uniform("u_prev_gate_interval_km")?,
-                prev_max_range_km: uniform("u_prev_max_range_km")?,
-            };
-
-            // Create placeholders for previous sweep textures
-            let prev_data_texture = create_r32f_texture(gl, 1, 1, &[0.0]);
-            let prev_azimuth_texture = create_r32f_texture(gl, 1, 1, &[0.0]);
-
-            gl.use_program(None);
-
-            Ok(Self {
-                program,
-                vao,
-                vbo,
-                data_texture,
-                lut_texture,
-                azimuth_texture,
-                prev_data_texture,
-                prev_azimuth_texture,
-                uniforms,
-                current: SweepState::default(),
-                prev: SweepState::default(),
-                cpu: CpuShadowData::default(),
-                prev_cpu: CpuShadowData::default(),
-                has_data: false,
-                value_min: 0.0,
-                value_range: 1.0,
-            })
-        }
-    }
-
-    /// Upload decoded radar data to GPU textures.
-    ///
-    /// `gate_values` contains raw u16 values cast to f32.
-    /// Sentinels: 0 = below threshold, 1 = range folded.
-    /// Physical value = (raw - offset) / scale.
-    #[allow(clippy::too_many_arguments)]
-    pub fn update_data(
-        &mut self,
-        gl: &glow::Context,
-        azimuths: &[f32],
-        gate_values: &[f32],
-        azimuth_count: u32,
-        gate_count: u32,
-        first_gate_km: f64,
-        gate_interval_km: f64,
-        max_range_km: f64,
-        offset: f32,
-        scale: f32,
-        radial_times: &[f64],
-    ) {
-        let t_total = web_time::Instant::now();
-
-        self.current.azimuth_count = azimuth_count;
-        self.current.gate_count = gate_count;
-        self.current.first_gate_km = first_gate_km;
-        self.current.gate_interval_km = gate_interval_km;
-        self.current.max_range_km = max_range_km;
-        self.current.data_offset = offset;
-        self.current.data_scale = scale;
-        self.has_data = azimuth_count > 0 && gate_count > 0;
-
-        // Keep CPU copies for inspector value lookup
-        let t_copy = web_time::Instant::now();
-        self.cpu.azimuths = azimuths.to_vec();
-        self.cpu.gate_values = gate_values.to_vec();
-        self.cpu.radial_times = radial_times.to_vec();
-        let copy_ms = t_copy.elapsed().as_secs_f64() * 1000.0;
-
-        if !self.has_data {
-            return;
-        }
-
-        let t_upload = web_time::Instant::now();
-        unsafe {
-            // Re-create data texture (gates x azimuths, R32F)
-            gl.delete_texture(self.data_texture);
-            self.data_texture =
-                create_r32f_texture(gl, gate_count as i32, azimuth_count as i32, gate_values);
-
-            // Re-create azimuth texture (Nx1, R32F)
-            gl.delete_texture(self.azimuth_texture);
-            self.azimuth_texture = create_r32f_texture(gl, azimuth_count as i32, 1, azimuths);
-        }
-        let upload_ms = t_upload.elapsed().as_secs_f64() * 1000.0;
-        let total_ms = t_total.elapsed().as_secs_f64() * 1000.0;
-
-        let expected_len = (azimuth_count * gate_count) as usize;
-        let actual_len = gate_values.len();
-        let az_first = azimuths.first().copied().unwrap_or(f32::NAN);
-        let az_last = azimuths.last().copied().unwrap_or(f32::NAN);
-        if actual_len != expected_len {
-            log::warn!(
-                "GPU update_data: dimension MISMATCH — {}x{} expects {} values, got {} (diff={})",
-                azimuth_count,
-                gate_count,
-                expected_len,
-                actual_len,
-                actual_len as i64 - expected_len as i64,
-            );
-        }
-        log::info!(
-            "GPU update_data: {}x{} (az x gates), azimuths=[{:.1}..{:.1}], vals={}, range {:.1}-{:.1} km, offset={} scale={}, {:.1}ms (copy: {:.1}ms, upload: {:.1}ms)",
-            azimuth_count,
-            gate_count,
-            az_first,
-            az_last,
-            actual_len,
-            first_gate_km,
-            max_range_km,
-            offset,
-            scale,
-            total_ms,
-            copy_ms,
-            upload_ms,
-        );
-    }
-
-    /// Build and upload a color lookup table for the given product.
-    pub fn update_color_table(&mut self, gl: &glow::Context, product_str: &str) {
-        let t_total = web_time::Instant::now();
-
-        let product = product_from_str(product_str);
-        let (min_val, max_val) = product_value_range(product);
-        self.value_min = min_val;
-        self.value_range = max_val - min_val;
-
-        let t_build = web_time::Instant::now();
-
-        // Build 1024-entry RGBA LUT (continuous gradient + GL_LINEAR = zero visible quantization)
-        let lut_size = 1024usize;
-        let lut_data = if matches!(product, Product::Reflectivity) {
-            // OKLab-interpolated reflectivity palette with alpha ramp
-            build_reflectivity_lut(min_val, max_val)
-        } else {
-            let color_scale = continuous_color_scale(product);
-            let mut data = Vec::with_capacity(lut_size * 4);
-            for i in 0..lut_size {
-                let t = i as f32 / (lut_size - 1) as f32;
-                let value = min_val + t * (max_val - min_val);
-                let color = color_scale.color(value);
-                let rgba = color.to_rgba8();
-                data.extend_from_slice(&rgba);
-            }
-            data
-        };
-        let build_ms = t_build.elapsed().as_secs_f64() * 1000.0;
-
-        let t_upload = web_time::Instant::now();
-        unsafe {
-            gl.delete_texture(self.lut_texture);
-            self.lut_texture = create_rgba8_texture(gl, lut_size as i32, 1, &lut_data);
-        }
-        let upload_ms = t_upload.elapsed().as_secs_f64() * 1000.0;
-        let total_ms = t_total.elapsed().as_secs_f64() * 1000.0;
-
-        log::info!(
-            "GPU update_color_table: {:?} ({:.1}..{:.1}), {:.1}ms (build: {:.1}ms, upload: {:.1}ms)",
-            product,
-            min_val,
-            max_val,
-            total_ms,
-            build_ms,
-            upload_ms,
-        );
-    }
-
-    /// Returns true if radar data has been uploaded.
-    pub fn has_data(&self) -> bool {
-        self.has_data
-    }
-
-    /// Maximum range of the currently loaded data in km.
-    pub fn max_range_km(&self) -> f64 {
-        self.current.max_range_km
-    }
-
-    // --- Accessors for globe radar renderer ---
-
-    pub fn gate_count(&self) -> u32 {
-        self.current.gate_count
-    }
-    pub fn azimuth_count(&self) -> u32 {
-        self.current.azimuth_count
-    }
-    pub fn first_gate_km(&self) -> f64 {
-        self.current.first_gate_km
-    }
-    pub fn gate_interval_km(&self) -> f64 {
-        self.current.gate_interval_km
-    }
-    pub fn value_min(&self) -> f32 {
-        self.value_min
-    }
-    pub fn value_range(&self) -> f32 {
-        self.value_range
-    }
-    pub fn data_offset(&self) -> f32 {
-        self.current.data_offset
-    }
-    pub fn data_scale(&self) -> f32 {
-        self.current.data_scale
-    }
-    pub fn data_texture(&self) -> glow::Texture {
-        self.data_texture
-    }
-    pub fn lut_texture(&self) -> glow::Texture {
-        self.lut_texture
-    }
-    pub fn azimuth_texture(&self) -> glow::Texture {
-        self.azimuth_texture
-    }
-
-    /// Clear all radar data (e.g. on site change).
-    pub fn clear_data(&mut self) {
-        self.has_data = false;
-        self.current.sweep_id = None;
-        self.prev.sweep_id = None;
-        self.cpu.azimuths.clear();
-        self.cpu.gate_values.clear();
-        self.cpu.radial_times.clear();
-    }
-
-    /// Clear only the previous-sweep identity so the shader composites against
-    /// black until a new previous sweep is loaded.
-    pub fn clear_previous_data(&mut self) {
-        self.prev.sweep_id = None;
-    }
-
-    /// Promote the current texture to the previous texture slot.
-    ///
-    /// Called at elevation transitions during live streaming so the just-completed
-    /// sweep becomes the background for compositing partial data on top.
-    pub fn promote_current_to_previous(&mut self, gl: &glow::Context) {
-        if !self.has_data || self.cpu.azimuths.is_empty() {
-            return;
-        }
-        self.update_previous_data(
-            gl,
-            &self.cpu.azimuths.clone(),
-            &self.cpu.gate_values.clone(),
-            self.current.azimuth_count,
-            self.current.gate_count,
-            self.current.first_gate_km,
-            self.current.gate_interval_km,
-            self.current.max_range_km,
-            self.current.data_offset,
-            self.current.data_scale,
-            self.current.sweep_id.clone(),
-            &self.cpu.radial_times.clone(),
-        );
-    }
-
-    /// Identity of the sweep currently loaded in the primary data texture.
-    pub fn current_sweep_id(&self) -> Option<&str> {
-        self.current.sweep_id.as_deref()
-    }
-
-    /// Identity of the sweep currently loaded in the previous data texture.
-    pub fn prev_sweep_id(&self) -> Option<&str> {
-        self.prev.sweep_id.as_deref()
-    }
-
-    /// Set the identity of the current sweep (called after `update_data`).
-    pub fn set_current_sweep_id(&mut self, id: Option<String>) {
-        self.current.sweep_id = id;
-    }
-
-    /// Upload decoded radar data to the *previous* texture slot for sweep
-    /// animation compositing. Stores per-sweep spatial metadata so the shader
-    /// can sample the previous texture with correct gate/range mapping even
-    /// when the current and previous sweeps have different dimensions.
-    #[allow(clippy::too_many_arguments)]
-    pub fn update_previous_data(
-        &mut self,
-        gl: &glow::Context,
-        azimuths: &[f32],
-        gate_values: &[f32],
-        azimuth_count: u32,
-        gate_count: u32,
-        first_gate_km: f64,
-        gate_interval_km: f64,
-        max_range_km: f64,
-        offset: f32,
-        scale: f32,
-        sweep_id: Option<String>,
-        radial_times: &[f64],
-    ) {
-        self.prev.data_offset = offset;
-        self.prev.data_scale = scale;
-        self.prev.azimuth_count = azimuth_count;
-        self.prev.gate_count = gate_count;
-        self.prev.first_gate_km = first_gate_km;
-        self.prev.gate_interval_km = gate_interval_km;
-        self.prev.max_range_km = max_range_km;
-        self.prev.sweep_id = sweep_id;
-        self.prev_cpu.gate_values = gate_values.to_vec();
-        self.prev_cpu.radial_times = radial_times.to_vec();
-
-        if azimuth_count == 0 || gate_count == 0 {
-            return;
-        }
-
-        unsafe {
-            gl.delete_texture(self.prev_data_texture);
-            self.prev_data_texture =
-                create_r32f_texture(gl, gate_count as i32, azimuth_count as i32, gate_values);
-
-            gl.delete_texture(self.prev_azimuth_texture);
-            self.prev_azimuth_texture = create_r32f_texture(gl, azimuth_count as i32, 1, azimuths);
-        }
-    }
-
-    /// Look up the raw data value at a given polar coordinate.
-    ///
-    /// When `sweep_params` is `Some((sweep_azimuth, sweep_start))`, determines
-    /// whether the queried position falls in the previous-sweep region and
-    /// returns the appropriate value. Pass `None` for non-animated lookups.
-    pub fn value_at_polar(
-        &self,
-        azimuth_deg: f32,
-        range_km: f64,
-        sweep_params: Option<(f32, f32)>,
-    ) -> Option<f32> {
-        if let Some((sweep_az, sweep_start)) = sweep_params {
-            let swept_arc = (sweep_az - sweep_start).rem_euclid(360.0);
-            let pixel_from_start = (azimuth_deg - sweep_start).rem_euclid(360.0);
-            if pixel_from_start >= swept_arc {
-                return self.prev_value_at_polar(azimuth_deg, range_km);
-            }
-        }
-        if !self.has_data || self.cpu.azimuths.is_empty() {
-            return None;
-        }
-
-        if range_km < self.current.first_gate_km || range_km >= self.current.max_range_km {
-            return None;
-        }
-
-        let az_idx = find_nearest_azimuth_index(
-            &self.cpu.azimuths,
-            self.current.azimuth_count as usize,
-            azimuth_deg,
-        )?;
-
-        let gate_count = self.current.gate_count as usize;
-        let gate_idx = ((range_km - self.current.first_gate_km) / self.current.gate_interval_km)
-            .floor() as usize;
-        if gate_idx >= gate_count {
-            return None;
-        }
-
-        let offset = az_idx * gate_count + gate_idx;
-        if offset >= self.cpu.gate_values.len() {
-            return None;
-        }
-
-        let raw = self.cpu.gate_values[offset];
-        if raw <= 1.0 {
-            return None;
-        }
-
-        if self.current.data_scale == 0.0 {
-            Some(raw)
-        } else {
-            Some((raw - self.current.data_offset) / self.current.data_scale)
-        }
-    }
-
-    /// Look up the radial collection timestamp (Unix seconds) at a given azimuth.
-    ///
-    /// When `sweep_params` is `Some((sweep_azimuth, sweep_start))`, determines
-    /// whether the queried position falls in the previous-sweep region and
-    /// returns the appropriate timestamp. Pass `None` for non-animated lookups.
-    pub fn collection_time_at_polar(
-        &self,
-        azimuth_deg: f32,
-        sweep_params: Option<(f32, f32)>,
-    ) -> Option<f64> {
-        if let Some((sweep_az, sweep_start)) = sweep_params {
-            let swept_arc = (sweep_az - sweep_start).rem_euclid(360.0);
-            let pixel_from_start = (azimuth_deg - sweep_start).rem_euclid(360.0);
-            if pixel_from_start >= swept_arc {
-                return self.prev_collection_time_at_polar(azimuth_deg);
-            }
-        }
-        if self.cpu.radial_times.is_empty() || self.cpu.azimuths.is_empty() {
-            return None;
-        }
-
-        let az_idx = find_nearest_azimuth_index(
-            &self.cpu.azimuths,
-            self.current.azimuth_count as usize,
-            azimuth_deg,
-        )?;
-
-        self.cpu.radial_times.get(az_idx).copied()
-    }
-
-    /// Look up value in the previous sweep's CPU data using evenly-spaced azimuth indexing.
-    fn prev_value_at_polar(&self, azimuth_deg: f32, range_km: f64) -> Option<f32> {
-        let az_count = self.prev.azimuth_count as usize;
-        let gate_count = self.prev.gate_count as usize;
-        if az_count == 0 || gate_count == 0 || self.prev_cpu.gate_values.is_empty() {
-            return None;
-        }
-
-        if range_km < self.prev.first_gate_km || range_km >= self.prev.max_range_km {
-            return None;
-        }
-
-        // Evenly-spaced azimuth indexing (same as GPU shader for prev sweep)
-        let az_idx = ((azimuth_deg * az_count as f32 / 360.0).round() as usize) % az_count;
-
-        let gate_idx =
-            ((range_km - self.prev.first_gate_km) / self.prev.gate_interval_km).floor() as usize;
-        if gate_idx >= gate_count {
-            return None;
-        }
-
-        let offset = az_idx * gate_count + gate_idx;
-        if offset >= self.prev_cpu.gate_values.len() {
-            return None;
-        }
-
-        let raw = self.prev_cpu.gate_values[offset];
-        if raw <= 1.0 {
-            return None;
-        }
-
-        if self.prev.data_scale == 0.0 {
-            Some(raw)
-        } else {
-            Some((raw - self.prev.data_offset) / self.prev.data_scale)
-        }
-    }
-
-    /// Look up collection time in the previous sweep's CPU data.
-    fn prev_collection_time_at_polar(&self, azimuth_deg: f32) -> Option<f64> {
-        let az_count = self.prev.azimuth_count as usize;
-        if az_count == 0 || self.prev_cpu.radial_times.is_empty() {
-            return None;
-        }
-
-        let az_idx = ((azimuth_deg * az_count as f32 / 360.0).round() as usize) % az_count;
-        self.prev_cpu.radial_times.get(az_idx).copied()
-    }
-
-    /// Render the radar data using the current GL context.
-    ///
-    /// Called from within an `egui_glow::CallbackFn`.
-    /// `radar_center` and `radar_radius` are in physical pixels (not points).
-    ///
-    /// egui_glow restores its own GL state after each paint callback,
-    /// so we don't need to save/restore state ourselves.
-    pub fn paint(
-        &self,
-        gl: &glow::Context,
-        radar_center: [f32; 2],
-        radar_radius: f32,
-        viewport_size: [f32; 2],
-        processing: &RenderProcessing,
-        sweep_info: Option<(f32, f32)>,
-    ) {
-        if !self.has_data {
-            return;
-        }
-
-        unsafe {
-            gl.use_program(Some(self.program));
-            gl.bind_vertex_array(Some(self.vao));
-
-            // Premultiplied alpha blending
-            gl.enable(glow::BLEND);
-            gl.blend_func_separate(
-                glow::ONE,
-                glow::ONE_MINUS_SRC_ALPHA,
-                glow::ONE,
-                glow::ONE_MINUS_SRC_ALPHA,
-            );
-            gl.disable(glow::SCISSOR_TEST);
-
-            // Bind textures to units
-            gl.active_texture(glow::TEXTURE0);
-            gl.bind_texture(glow::TEXTURE_2D, Some(self.data_texture));
-            gl.active_texture(glow::TEXTURE1);
-            gl.bind_texture(glow::TEXTURE_2D, Some(self.lut_texture));
-            gl.active_texture(glow::TEXTURE2);
-            gl.bind_texture(glow::TEXTURE_2D, Some(self.azimuth_texture));
-            gl.active_texture(glow::TEXTURE3);
-            gl.bind_texture(glow::TEXTURE_2D, Some(self.prev_data_texture));
-            gl.active_texture(glow::TEXTURE4);
-            gl.bind_texture(glow::TEXTURE_2D, Some(self.prev_azimuth_texture));
-
-            // Set uniforms
-            gl.uniform_2_f32(
-                Some(&self.uniforms.radar_center),
-                radar_center[0],
-                radar_center[1],
-            );
-            gl.uniform_1_f32(Some(&self.uniforms.radar_radius), radar_radius);
-            gl.uniform_1_f32(
-                Some(&self.uniforms.gate_count),
-                self.current.gate_count as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.azimuth_count),
-                self.current.azimuth_count as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.first_gate_km),
-                self.current.first_gate_km as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.gate_interval_km),
-                self.current.gate_interval_km as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.max_range_km),
-                self.current.max_range_km as f32,
-            );
-            gl.uniform_1_f32(Some(&self.uniforms.value_min), self.value_min);
-            gl.uniform_1_f32(Some(&self.uniforms.value_range), self.value_range);
-            gl.uniform_2_f32(
-                Some(&self.uniforms.viewport_size),
-                viewport_size[0],
-                viewport_size[1],
-            );
-
-            // Processing uniforms
-            let interp_mode = match processing.interpolation {
-                crate::state::InterpolationMode::Nearest => 0,
-                crate::state::InterpolationMode::Bilinear => 1,
-            };
-            gl.uniform_1_i32(Some(&self.uniforms.interpolation), interp_mode);
-            gl.uniform_1_i32(
-                Some(&self.uniforms.despeckle_enabled),
-                processing.despeckle_enabled as i32,
-            );
-            gl.uniform_1_i32(
-                Some(&self.uniforms.despeckle_threshold),
-                processing.despeckle_threshold as i32,
-            );
-            gl.uniform_1_f32(Some(&self.uniforms.opacity), processing.opacity);
-            gl.uniform_1_i32(
-                Some(&self.uniforms.data_age_indicator),
-                processing.data_age_indicator as i32,
-            );
-
-            // Raw-to-physical conversion
-            gl.uniform_1_f32(Some(&self.uniforms.offset), self.current.data_offset);
-            gl.uniform_1_f32(Some(&self.uniforms.scale), self.current.data_scale);
-
-            // Sweep animation uniforms — enable even without prev data; the 1x1
-            // placeholder texture returns 0.0 (below-threshold sentinel -> transparent),
-            // so the first sweep progressively reveals against a blank background.
-            let sweep_on = sweep_info.is_some();
-            gl.uniform_1_i32(Some(&self.uniforms.sweep_enabled), sweep_on as i32);
-            let (sweep_az, sweep_start) = sweep_info.unwrap_or((0.0, 0.0));
-            gl.uniform_1_f32(Some(&self.uniforms.sweep_azimuth), sweep_az);
-            gl.uniform_1_f32(Some(&self.uniforms.sweep_start), sweep_start);
-            gl.uniform_1_f32(Some(&self.uniforms.prev_offset), self.prev.data_offset);
-            gl.uniform_1_f32(Some(&self.uniforms.prev_scale), self.prev.data_scale);
-            gl.uniform_1_f32(
-                Some(&self.uniforms.prev_gate_count),
-                self.prev.gate_count as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.prev_azimuth_count),
-                self.prev.azimuth_count as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.prev_first_gate_km),
-                self.prev.first_gate_km as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.prev_gate_interval_km),
-                self.prev.gate_interval_km as f32,
-            );
-            gl.uniform_1_f32(
-                Some(&self.uniforms.prev_max_range_km),
-                self.prev.max_range_km as f32,
-            );
-
-            // Draw fullscreen quad
-            gl.draw_arrays(glow::TRIANGLES, 0, 6);
-
-            // Unbind our resources so we don't interfere with egui
-            gl.bind_vertex_array(None);
-            gl.use_program(None);
-            gl.active_texture(glow::TEXTURE0);
-        }
-    }
-
-    /// Detect storm cells from the current CPU-side data.
-    ///
-    /// Returns lightweight cell info for rendering on the canvas.
-    /// Uses nexrad-process connected-component analysis on the reflectivity data.
-    pub fn detect_storm_cells(
-        &self,
-        radar_lat: f64,
-        radar_lon: f64,
-        threshold_dbz: f32,
-    ) -> Vec<crate::state::StormCellInfo> {
-        if !self.has_data || self.cpu.azimuths.is_empty() {
-            return Vec::new();
-        }
-
-        let t_total = web_time::Instant::now();
-
-        let az_count = self.current.azimuth_count as usize;
-        let gate_count = self.current.gate_count as usize;
-
-        // Compute azimuth spacing
-        let az_spacing = if az_count > 1 {
-            360.0 / az_count as f32
-        } else {
-            1.0
-        };
-
-        // Build a SweepField from the CPU data
-        let t_field = web_time::Instant::now();
-        let mut field = nexrad_model::data::SweepField::new_empty(
-            "Reflectivity",
-            "dBZ",
-            0.5, // elevation doesn't matter for 2D detection
-            self.cpu.azimuths.clone(),
-            az_spacing,
-            self.current.first_gate_km,
-            self.current.gate_interval_km,
-            gate_count,
-        );
-
-        // Populate the field with our gate values (convert raw -> physical)
-        let mut valid_gates = 0u32;
-        for az_idx in 0..az_count {
-            let row_start = az_idx * gate_count;
-            for g in 0..gate_count {
-                let raw = self.cpu.gate_values[row_start + g];
-                // Raw sentinels: 0 = below threshold, 1 = range folded
-                if raw > 1.0 {
-                    let physical = if self.current.data_scale == 0.0 {
-                        raw
-                    } else {
-                        (raw - self.current.data_offset) / self.current.data_scale
-                    };
-                    field.set(az_idx, g, physical, nexrad_model::data::GateStatus::Valid);
-                    valid_gates += 1;
-                }
-                // new_empty defaults to NoData, so we only set Valid gates
-            }
-        }
-        let field_ms = t_field.elapsed().as_secs_f64() * 1000.0;
-
-        // Build coordinate system from site location
-        use nexrad_model::geo::RadarCoordinateSystem;
-        use nexrad_model::meta::Site;
-        use nexrad_process::detection::StormCellDetector;
-
-        let site = Site::new(
-            *b"SITE",
-            radar_lat as f32,
-            radar_lon as f32,
-            0, // altitude (not critical for 2D detection)
-            0, // tower height
-        );
-        let coord_system = RadarCoordinateSystem::new(&site);
-
-        // Run detection
-        let t_detect = web_time::Instant::now();
-        let detector: StormCellDetector = match StormCellDetector::new(threshold_dbz, 10) {
-            Ok(d) => d,
-            Err(_) => return Vec::new(),
-        };
-
-        let cells: Vec<nexrad_process::detection::StormCell> =
-            match detector.detect(&field, &coord_system) {
-                Ok(c) => c,
-                Err(e) => {
-                    log::warn!("Storm cell detection failed: {}", e);
-                    return Vec::new();
-                }
-            };
-        let detect_ms = t_detect.elapsed().as_secs_f64() * 1000.0;
-
-        // Convert to lightweight info, filtering out small noise cells
-        let t_convert = web_time::Instant::now();
-        const MIN_AREA_KM2: f64 = 5.0;
-
-        let result: Vec<_> = cells
-            .iter()
-            .filter(|cell| cell.area_km2() >= MIN_AREA_KM2)
-            .map(|cell| {
-                let centroid = cell.centroid();
-                let bounds = cell.bounds();
-                crate::state::StormCellInfo {
-                    lat: centroid.latitude,
-                    lon: centroid.longitude,
-                    max_dbz: cell.max_reflectivity_dbz(),
-                    area_km2: cell.area_km2() as f32,
-                    bounds: (
-                        bounds.min_latitude(),
-                        bounds.min_longitude(),
-                        bounds.max_latitude(),
-                        bounds.max_longitude(),
-                    ),
-                }
-            })
-            .collect();
-        let convert_ms = t_convert.elapsed().as_secs_f64() * 1000.0;
-        let total_ms = t_total.elapsed().as_secs_f64() * 1000.0;
-
-        log::info!(
-            "detect_storm_cells: {}x{} grid, {} valid gates, {} raw cells, {} after filter (>={:.0} km2), {:.1}ms (field: {:.1}ms, detect: {:.1}ms, convert: {:.1}ms)",
-            az_count,
-            gate_count,
-            valid_gates,
-            cells.len(),
-            result.len(),
-            MIN_AREA_KM2,
-            total_ms,
-            field_ms,
-            detect_ms,
-            convert_ms,
-        );
-
-        result
-    }
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-/// Cast an `&[f32]` to `&[u8]` for GL upload.
-fn bytemuck_cast_slice(data: &[f32]) -> &[u8] {
-    unsafe { std::slice::from_raw_parts(data.as_ptr() as *const u8, data.len() * 4) }
-}
-
-/// Create an R32F texture with the given dimensions and data.
-unsafe fn create_r32f_texture(
-    gl: &glow::Context,
-    width: i32,
-    height: i32,
-    data: &[f32],
-) -> glow::Texture {
-    let texture = gl.create_texture().expect("Cannot create texture");
-    gl.bind_texture(glow::TEXTURE_2D, Some(texture));
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_MIN_FILTER,
-        glow::NEAREST as i32,
-    );
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_MAG_FILTER,
-        glow::NEAREST as i32,
-    );
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_WRAP_S,
-        glow::CLAMP_TO_EDGE as i32,
-    );
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_WRAP_T,
-        glow::CLAMP_TO_EDGE as i32,
-    );
-    gl.tex_image_2d(
-        glow::TEXTURE_2D,
-        0,
-        glow::R32F as i32,
-        width,
-        height,
-        0,
-        glow::RED,
-        glow::FLOAT,
-        glow::PixelUnpackData::Slice(Some(bytemuck_cast_slice(data))),
-    );
-    gl.bind_texture(glow::TEXTURE_2D, None);
-    texture
-}
-
-/// Create an RGBA8 texture with the given dimensions and data.
-unsafe fn create_rgba8_texture(
-    gl: &glow::Context,
-    width: i32,
-    height: i32,
-    data: &[u8],
-) -> glow::Texture {
-    let texture = gl.create_texture().expect("Cannot create texture");
-    gl.bind_texture(glow::TEXTURE_2D, Some(texture));
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_MIN_FILTER,
-        glow::LINEAR as i32,
-    );
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_MAG_FILTER,
-        glow::LINEAR as i32,
-    );
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_WRAP_S,
-        glow::CLAMP_TO_EDGE as i32,
-    );
-    gl.tex_parameter_i32(
-        glow::TEXTURE_2D,
-        glow::TEXTURE_WRAP_T,
-        glow::CLAMP_TO_EDGE as i32,
-    );
-    gl.tex_image_2d(
-        glow::TEXTURE_2D,
-        0,
-        glow::RGBA as i32,
-        width,
-        height,
-        0,
-        glow::RGBA,
-        glow::UNSIGNED_BYTE,
-        glow::PixelUnpackData::Slice(Some(data)),
-    );
-    gl.bind_texture(glow::TEXTURE_2D, None);
-    texture
-}
