+++ src/nexrad/worker_api/ingest.rs
@@ -1,221 +1,6 @@
-//! WASM exports for the Web Worker.
-//!
-//! These functions are called from worker.js to perform heavy data operations
-//! (ingest, render) in a background thread, keeping the main UI responsive.
+//! WASM exports for ingest operations (full archive and per-chunk streaming).
 
-use crate::data::indexeddb::IndexedDbRecordStore;
-use crate::data::keys::*;
-use serde::{Deserialize, Serialize};
-use wasm_bindgen::JsValue;
-
-// ---------------------------------------------------------------------------
-// Typed input param structs — deserialized from JS objects via serde-wasm-bindgen
-// ---------------------------------------------------------------------------
-
-/// Parameters for `worker_ingest`. The `data` ArrayBuffer is extracted separately.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct IngestParams {
-    site_id: String,
-    timestamp_secs: f64,
-    #[serde(default)]
-    file_name: String,
-}
-
-/// Parameters for `worker_render`.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct RenderParams {
-    scan_key: String,
-    elevation_number: u8,
-    #[serde(default = "default_product")]
-    product: String,
-}
-
-/// Parameters for `worker_render_volume`.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct RenderVolumeParams {
-    scan_key: String,
-    #[serde(default = "default_product")]
-    product: String,
-    elevation_numbers: Vec<u8>,
-}
-
-/// Parameters for `worker_ingest_chunk`. The `data` ArrayBuffer is extracted separately.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct IngestChunkParams {
-    site_id: String,
-    timestamp_secs: f64,
-    #[serde(default)]
-    chunk_index: u32,
-    #[serde(default)]
-    is_start: bool,
-    #[serde(default)]
-    is_end: bool,
-    #[serde(default)]
-    file_name: String,
-}
-
-fn default_product() -> String {
-    "reflectivity".to_string()
-}
-
-/// Extract the `data` ArrayBuffer field from a JS object as `Vec<u8>`.
-fn extract_data_bytes(obj: &JsValue) -> Result<Vec<u8>, JsValue> {
-    let val = js_sys::Reflect::get(obj, &"data".into())
-        .map_err(|e| JsValue::from_str(&format!("Missing data: {:?}", e)))?;
-    Ok(js_sys::Uint8Array::new(&val).to_vec())
-}
-
-// ---------------------------------------------------------------------------
-// Typed response structs — serialized to JS objects via serde-wasm-bindgen
-// ---------------------------------------------------------------------------
-
-/// Response from `worker_ingest`.
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct IngestResponse<'a> {
-    records_stored: u32,
-    scan_key: String,
-    elevation_numbers: &'a [u8],
-    total_ms: f64,
-    split_ms: f64,
-    decompress_ms: f64,
-    decode_ms: f64,
-    extract_ms: f64,
-    store_ms: f64,
-    index_ms: f64,
-    sweeps: &'a [SweepMeta],
-    #[serde(skip_serializing_if = "Option::is_none")]
-    vcp: Option<&'a ExtractedVcp>,
-}
-
-/// Scalar fields of the response from `worker_render`.
-/// ArrayBuffer fields (azimuths, gateValues, radialTimes) are set separately.
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct RenderResponse {
-    azimuth_count: u32,
-    gate_count: u32,
-    first_gate_range_km: f64,
-    gate_interval_km: f64,
-    max_range_km: f64,
-    product: String,
-    radial_count: u32,
-    scale: f64,
-    offset: f64,
-    mean_elevation: f64,
-    sweep_start_secs: f64,
-    sweep_end_secs: f64,
-    fetch_ms: f64,
-    deser_ms: f64,
-    total_ms: f64,
-    marshal_ms: f64,
-}
-
-/// Response from `worker_ingest_chunk`.
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ChunkIngestResponse {
-    chunk_index: u32,
-    radials_decoded: u32,
-    sweeps_stored: u32,
-    scan_key: String,
-    is_end: bool,
-    total_ms: f64,
-    sweeps: Vec<SweepMeta>,
-    elevations_completed: Vec<u8>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    vcp: Option<ExtractedVcp>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    chunk_min_time_secs: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    chunk_max_time_secs: Option<f64>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    chunk_elev_spans: Vec<(u8, f64, f64, u32)>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    chunk_elev_az_ranges: Vec<(u8, f32, f32)>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    volume_header_time_secs: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_radial_azimuth: Option<f32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    last_radial_time_secs: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    current_elevation: Option<u8>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    current_elevation_radials: Option<u32>,
-}
-
-/// Per-sweep metadata in the volume render response.
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct VolumeRenderSweepMeta {
-    elevation_deg: f64,
-    azimuth_count: u32,
-    gate_count: u32,
-    first_gate_km: f64,
-    gate_interval_km: f64,
-    max_range_km: f64,
-    data_offset: u32,
-    scale: f64,
-    offset: f64,
-}
-
-/// Scalar fields of the volume render response.
-/// The `buffer` ArrayBuffer is set separately for zero-copy transfer.
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct VolumeRenderResponse {
-    sweep_count: u32,
-    word_size: u8,
-    sweep_meta: Vec<VolumeRenderSweepMeta>,
-    product: String,
-    total_ms: f64,
-}
-
-// ---------------------------------------------------------------------------
-// Worker-side cached IDB connection
-// ---------------------------------------------------------------------------
-// WASM is single-threaded so thread_local! is safe. We keep a single
-// IndexedDbRecordStore alive for the lifetime of the worker so that
-// subsequent ingest/render calls reuse the already-open IDB connection
-// instead of paying the ~60ms open+list overhead every time.
-
-thread_local! {
-    static WORKER_IDB: std::cell::RefCell<Option<IndexedDbRecordStore>> =
-        const { std::cell::RefCell::new(None) };
-    static WORKER_LOGGER_INIT: std::cell::Cell<bool> = const { std::cell::Cell::new(false) };
-}
-
-/// Initialize the log crate in the worker context (once).
-fn init_logger() {
-    WORKER_LOGGER_INIT.with(|init| {
-        if !init.get() {
-            eframe::WebLogger::init(log::LevelFilter::Debug).ok();
-            init.set(true);
-        }
-    });
-}
-
-/// Get (or lazily open) the shared worker IDB store.
-async fn idb_store() -> Result<IndexedDbRecordStore, wasm_bindgen::JsValue> {
-    let existing = WORKER_IDB.with(|cell| cell.borrow().clone());
-    if let Some(store) = existing {
-        return Ok(store);
-    }
-    let store = IndexedDbRecordStore::new();
-    store
-        .open()
-        .await
-        .map_err(|e| wasm_bindgen::JsValue::from_str(&format!("Failed to open IDB: {}", e)))?;
-    WORKER_IDB.with(|cell| {
-        *cell.borrow_mut() = Some(store.clone());
-    });
-    Ok(store)
-}
+use super::*;
 
 /// Ingest a raw NEXRAD archive file: split into LDM records, probe for elevation
 /// metadata, store in IndexedDB, and return metadata.
@@ -392,463 +177,6 @@ pub fn worker_ingest(params: wasm_bindgen::JsValue) -> js_sys::Promise {
     })
 }
 
-/// Render a specific elevation from pre-computed sweep data in IndexedDB.
-///
-/// Called from the Web Worker via worker.js. Fetches a single pre-computed
-/// sweep blob and returns the data for GPU upload — no decoding needed.
-///
-/// Parameters (JS object): `{ scanKey: string, elevationNumber: number, product: string }`
-/// Returns (JS object): `{ azimuths: Float32Array, gateValues: Float32Array, azimuthCount, gateCount, ... }`
-#[wasm_bindgen::prelude::wasm_bindgen]
-pub fn worker_render(params: wasm_bindgen::JsValue) -> js_sys::Promise {
-    init_logger();
-    wasm_bindgen_futures::future_to_promise(async move {
-        let t_total = web_time::Instant::now();
-
-        let p: RenderParams = serde_wasm_bindgen::from_value(params)
-            .map_err(|e| JsValue::from_str(&format!("Invalid render params: {}", e)))?;
-        let scan_key_str = p.scan_key;
-        let elevation_number = p.elevation_number;
-        let product_str = p.product;
-
-        let scan_key = ScanKey::from_storage_key(&scan_key_str)
-            .ok_or_else(|| wasm_bindgen::JsValue::from_str("Invalid scanKey format"))?;
-
-        let store = idb_store().await?;
-
-        // Fetch raw IDB ArrayBuffer (no Rust-side copy)
-        let t_fetch = web_time::Instant::now();
-        let sweep_key = SweepDataKey::new(scan_key, elevation_number, &product_str);
-        let blob_buffer = store
-            .get_sweep_as_js(&sweep_key.to_storage_key())
-            .await
-            .map_err(|e| wasm_bindgen::JsValue::from_str(&format!("Failed to fetch sweep: {}", e)))?
-            .ok_or_else(|| {
-                wasm_bindgen::JsValue::from_str(&format!(
-                    "No pre-computed sweep for elev={} product={}",
-                    elevation_number, product_str
-                ))
-            })?;
-        let fetch_ms = t_fetch.elapsed().as_secs_f64() * 1000.0;
-        let blob_len = blob_buffer.byte_length();
-
-        // Parse header only (72 bytes) — no array allocations
-        let t_deser = web_time::Instant::now();
-        let header_bytes = {
-            let view = js_sys::Uint8Array::new_with_byte_offset_and_length(&blob_buffer, 0, 72);
-            let mut buf = [0u8; 72];
-            view.copy_to(&mut buf);
-            buf
-        };
-        let header = parse_sweep_header(&header_bytes).map_err(|e| {
-            wasm_bindgen::JsValue::from_str(&format!("Failed to parse sweep header: {}", e))
-        })?;
-
-        // Validate full blob size
-        let az = header.azimuth_count as usize;
-        let gc = header.gate_count as usize;
-        let ws = header.data_word_size as usize;
-        let expected = header.gate_values_offset as usize + az * gc * ws;
-        if (blob_len as usize) < expected {
-            return Err(wasm_bindgen::JsValue::from_str(&format!(
-                "Sweep blob too small: {} < {} expected",
-                blob_len, expected
-            )));
-        }
-        let deser_ms = t_deser.elapsed().as_secs_f64() * 1000.0;
-
-        // Marshal: create typed array views over raw IDB ArrayBuffer
-        let t_marshal = web_time::Instant::now();
-
-        let az_view = js_sys::Float32Array::new_with_byte_offset_and_length(
-            &blob_buffer,
-            header.azimuths_offset,
-            header.azimuth_count,
-        );
-        let az_buf = az_view.slice(0, header.azimuth_count).buffer();
-
-        // Extract radial_times if present (format version >= 1)
-        let rt_buf = if header.radial_times_offset > 0 {
-            let rt_view = js_sys::Float64Array::new_with_byte_offset_and_length(
-                &blob_buffer,
-                header.radial_times_offset as u32,
-                header.azimuth_count,
-            );
-            Some(rt_view.slice(0, header.azimuth_count).buffer())
-        } else {
-            None
-        };
-
-        // Convert native-width gate values to f32 for GPU upload
-        let gate_count_total = header.azimuth_count * header.gate_count;
-        let val_buf = if header.data_word_size == 1 {
-            let u8_view = js_sys::Uint8Array::new_with_byte_offset_and_length(
-                &blob_buffer,
-                header.gate_values_offset,
-                gate_count_total,
-            );
-            js_sys::Float32Array::new(&u8_view).buffer()
-        } else {
-            let u16_view = js_sys::Uint16Array::new_with_byte_offset_and_length(
-                &blob_buffer,
-                header.gate_values_offset,
-                gate_count_total,
-            );
-            js_sys::Float32Array::new(&u16_view).buffer()
-        };
-
-        let marshal_ms = t_marshal.elapsed().as_secs_f64() * 1000.0;
-        let total_ms = t_total.elapsed().as_secs_f64() * 1000.0;
-
-        log::info!(
-            "render: elev={} {} {}x{} ({:.1}KB) in {:.1}ms | fetch {:.1} | deser {:.1} | marshal {:.1}",
-            elevation_number, product_str,
-            header.azimuth_count, header.gate_count,
-            blob_len as f64 / 1024.0,
-            total_ms, fetch_ms, deser_ms, marshal_ms,
-        );
-
-        // Serialize scalar fields, then attach ArrayBuffer fields separately
-        let response = RenderResponse {
-            azimuth_count: header.azimuth_count,
-            gate_count: header.gate_count,
-            first_gate_range_km: header.first_gate_range_km,
-            gate_interval_km: header.gate_interval_km,
-            max_range_km: header.max_range_km,
-            product: product_str,
-            radial_count: header.radial_count,
-            scale: header.scale as f64,
-            offset: header.offset as f64,
-            mean_elevation: header.mean_elevation as f64,
-            sweep_start_secs: header.sweep_start_secs,
-            sweep_end_secs: header.sweep_end_secs,
-            fetch_ms,
-            deser_ms,
-            total_ms,
-            marshal_ms,
-        };
-        let result = serde_wasm_bindgen::to_value(&response)
-            .map_err(|e| JsValue::from_str(&format!("Failed to serialize response: {}", e)))?;
-        // ArrayBuffer fields must be set directly (not serializable via serde)
-        js_sys::Reflect::set(&result, &"azimuths".into(), &az_buf).ok();
-        js_sys::Reflect::set(&result, &"gateValues".into(), &val_buf).ok();
-        if let Some(rt) = rt_buf {
-            js_sys::Reflect::set(&result, &"radialTimes".into(), &rt).ok();
-        }
-        Ok(result)
-    })
-}
-
-// ---------------------------------------------------------------------------
-// Live (partial sweep) render from in-memory ChunkAccumulator
-// ---------------------------------------------------------------------------
-
-/// Parameters for `worker_render_live`.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct RenderLiveParams {
-    #[serde(default = "default_product")]
-    product: String,
-    #[serde(default)]
-    elevation_number: Option<u8>,
-}
-
-/// Render the current partial sweep from the in-memory ChunkAccumulator.
-///
-/// This reads directly from memory (no IDB), so it's very fast (~1ms).
-/// Returns the same RenderResponse shape as `worker_render`.
-///
-/// Parameters (JS object): `{ product: string, elevationNumber?: number }`
-#[wasm_bindgen::prelude::wasm_bindgen]
-pub fn worker_render_live(params: wasm_bindgen::JsValue) -> Result<JsValue, JsValue> {
-    use crate::nexrad::record_decode::extract_sweep_data_from_sorted;
-    use nexrad_render::Product;
-
-    let t_total = web_time::Instant::now();
-
-    let p: RenderLiveParams = serde_wasm_bindgen::from_value(params)
-        .map_err(|e| JsValue::from_str(&format!("Invalid render_live params: {}", e)))?;
-
-    let product = match p.product.as_str() {
-        "velocity" => Product::Velocity,
-        "spectrum_width" => Product::SpectrumWidth,
-        "differential_reflectivity" => Product::DifferentialReflectivity,
-        "differential_phase" => Product::DifferentialPhase,
-        "correlation_coefficient" => Product::CorrelationCoefficient,
-        "clutter_filter_power" => Product::ClutterFilterPower,
-        _ => Product::Reflectivity,
-    };
-
-    CHUNK_ACCUM.with(|cell| {
-        let borrow = cell.borrow();
-        let accum = borrow
-            .as_ref()
-            .ok_or_else(|| JsValue::from_str("No chunk accumulator active"))?;
-
-        let target_elev = p
-            .elevation_number
-            .or(accum.last_elevation_number)
-            .ok_or_else(|| JsValue::from_str("No elevation available in accumulator"))?;
-
-        // Filter and sort radials for the target elevation
-        let mut sorted: Vec<&::nexrad::model::data::Radial> = accum
-            .all_radials
-            .iter()
-            .filter(|r| r.elevation_number() == target_elev)
-            .collect();
-
-        if sorted.is_empty() {
-            return Err(JsValue::from_str(&format!(
-                "No radials for elevation {} in accumulator",
-                target_elev
-            )));
-        }
-
-        sorted.sort_by(|a, b| {
-            a.azimuth_angle_degrees()
-                .partial_cmp(&b.azimuth_angle_degrees())
-                .unwrap_or(std::cmp::Ordering::Equal)
-        });
-
-        let sweep = extract_sweep_data_from_sorted(&sorted, product).ok_or_else(|| {
-            JsValue::from_str(&format!(
-                "No {} data for elevation {} in accumulator",
-                p.product, target_elev
-            ))
-        })?;
-
-        // Marshal PrecomputedSweep into the same JS response format as worker_render
-        let t_marshal = web_time::Instant::now();
-
-        // Convert gate values to f32 array
-        let gate_values_f32: Vec<f32> = match &sweep.gate_values {
-            crate::data::keys::GateValues::U8(v) => v.iter().map(|&x| x as f32).collect(),
-            crate::data::keys::GateValues::U16(v) => v.iter().map(|&x| x as f32).collect(),
-        };
-
-        let az_array = js_sys::Float32Array::from(sweep.azimuths.as_slice());
-        let az_buf = az_array.buffer();
-
-        let val_array = js_sys::Float32Array::from(gate_values_f32.as_slice());
-        let val_buf = val_array.buffer();
-
-        let rt_array = js_sys::Float64Array::from(sweep.radial_times.as_slice());
-        let rt_buf = rt_array.buffer();
-
-        let marshal_ms = t_marshal.elapsed().as_secs_f64() * 1000.0;
-        let total_ms = t_total.elapsed().as_secs_f64() * 1000.0;
-
-        let accum_total = accum.all_radials.len();
-        let elev_radials = sorted.len();
-        let product_radials = sweep.azimuth_count;
-        let expected_values = sweep.azimuth_count as usize * sweep.gate_count as usize;
-        let actual_values = gate_values_f32.len();
-        log::info!(
-            "render_live: elev={} {} {}x{} accum_total={} elev_radials={} product_radials={} vals={}/{} az=[{:.1}..{:.1}] offset={} scale={} in {:.1}ms (marshal: {:.1}ms)",
-            target_elev,
-            p.product,
-            sweep.azimuth_count,
-            sweep.gate_count,
-            accum_total,
-            elev_radials,
-            product_radials,
-            actual_values,
-            expected_values,
-            sweep.azimuths.first().copied().unwrap_or(f32::NAN),
-            sweep.azimuths.last().copied().unwrap_or(f32::NAN),
-            sweep.offset,
-            sweep.scale,
-            total_ms,
-            marshal_ms,
-        );
-
-        let response = RenderResponse {
-            azimuth_count: sweep.azimuth_count,
-            gate_count: sweep.gate_count,
-            first_gate_range_km: sweep.first_gate_range_km,
-            gate_interval_km: sweep.gate_interval_km,
-            max_range_km: sweep.max_range_km,
-            product: p.product,
-            radial_count: sweep.radial_count,
-            scale: sweep.scale as f64,
-            offset: sweep.offset as f64,
-            mean_elevation: sweep.mean_elevation as f64,
-            sweep_start_secs: sweep.sweep_start_secs,
-            sweep_end_secs: sweep.sweep_end_secs,
-            fetch_ms: 0.0,
-            deser_ms: 0.0,
-            total_ms,
-            marshal_ms,
-        };
-        let result = serde_wasm_bindgen::to_value(&response)
-            .map_err(|e| JsValue::from_str(&format!("Failed to serialize response: {}", e)))?;
-        js_sys::Reflect::set(&result, &"azimuths".into(), &az_buf).ok();
-        js_sys::Reflect::set(&result, &"gateValues".into(), &val_buf).ok();
-        js_sys::Reflect::set(&result, &"radialTimes".into(), &rt_buf).ok();
-        Ok(result)
-    })
-}
-
-// ---------------------------------------------------------------------------
-// Volume render (all elevations packed for ray marching)
-// ---------------------------------------------------------------------------
-
-/// Render all elevations for a scan, packing raw gate data into a single buffer
-/// for volumetric ray-march rendering on the GPU.
-///
-/// Parameters (JS object): `{ scanKey: string, product: string, elevationNumbers: number[] }`
-/// Returns (JS object): `{ buffer: ArrayBuffer, sweepMeta: [...], product, totalMs }`
-#[wasm_bindgen::prelude::wasm_bindgen]
-pub fn worker_render_volume(params: wasm_bindgen::JsValue) -> js_sys::Promise {
-    init_logger();
-    wasm_bindgen_futures::future_to_promise(async move {
-        let t_total = web_time::Instant::now();
-
-        let p: RenderVolumeParams = serde_wasm_bindgen::from_value(params)
-            .map_err(|e| JsValue::from_str(&format!("Invalid render_volume params: {}", e)))?;
-        let scan_key_str = p.scan_key;
-        let product_str = p.product;
-        let elevation_numbers = p.elevation_numbers;
-
-        let scan_key = ScanKey::from_storage_key(&scan_key_str)
-            .ok_or_else(|| wasm_bindgen::JsValue::from_str("Invalid scanKey format"))?;
-
-        let store = idb_store().await?;
-
-        // Collect all sweep data into a packed buffer.
-        // We keep native word size when all sweeps are u8 to halve transfer cost.
-        // Only widen to u16 when at least one sweep has u16 data.
-        let mut packed_data: Vec<u8> = Vec::new();
-        let mut sweep_meta_vec: Vec<VolumeRenderSweepMeta> = Vec::new();
-        let mut data_offset: u32 = 0; // offset in values (not bytes)
-        let mut has_u16 = false;
-
-        // First pass: read all sweep blobs and headers, determine word size
-        struct SweepBlob {
-            blob_buffer: js_sys::ArrayBuffer,
-            header: SweepHeader,
-            total_values: usize,
-        }
-        let mut sweep_blobs: Vec<SweepBlob> = Vec::new();
-
-        for &elev_num in &elevation_numbers {
-            let sweep_key = SweepDataKey::new(scan_key.clone(), elev_num, &product_str);
-            let blob_buffer = match store.get_sweep_as_js(&sweep_key.to_storage_key()).await {
-                Ok(Some(buf)) => buf,
-                _ => continue, // skip missing elevations
-            };
-
-            let header_bytes = {
-                let view = js_sys::Uint8Array::new_with_byte_offset_and_length(&blob_buffer, 0, 72);
-                let mut buf = [0u8; 72];
-                view.copy_to(&mut buf);
-                buf
-            };
-            let header = match parse_sweep_header(&header_bytes) {
-                Ok(h) => h,
-                Err(_) => continue,
-            };
-
-            if header.data_word_size != 1 {
-                has_u16 = true;
-            }
-
-            let total_values = header.azimuth_count as usize * header.gate_count as usize;
-            sweep_blobs.push(SweepBlob {
-                blob_buffer,
-                header,
-                total_values,
-            });
-        }
-
-        // Second pass: pack data using native word size when all u8,
-        // widening to u16 only when mixed.
-        let word_size: u8 = if has_u16 { 2 } else { 1 };
-
-        for sb in &sweep_blobs {
-            let header = &sb.header;
-            let total_values = sb.total_values;
-
-            if word_size == 1 {
-                // All sweeps are u8: copy raw bytes, no widening
-                let u8_view = js_sys::Uint8Array::new_with_byte_offset_and_length(
-                    &sb.blob_buffer,
-                    header.gate_values_offset,
-                    total_values as u32,
-                );
-                let prev_len = packed_data.len();
-                packed_data.resize(prev_len + total_values, 0);
-                u8_view.copy_to(&mut packed_data[prev_len..]);
-            } else if header.data_word_size == 1 {
-                // Mixed volume: widen this u8 sweep to u16
-                let u8_view = js_sys::Uint8Array::new_with_byte_offset_and_length(
-                    &sb.blob_buffer,
-                    header.gate_values_offset,
-                    total_values as u32,
-                );
-                let mut tmp = vec![0u8; total_values];
-                u8_view.copy_to(&mut tmp);
-                for &val in &tmp {
-                    packed_data.extend_from_slice(&(val as u16).to_le_bytes());
-                }
-            } else {
-                // Native u16: copy raw bytes directly
-                let u8_view = js_sys::Uint8Array::new_with_byte_offset_and_length(
-                    &sb.blob_buffer,
-                    header.gate_values_offset,
-                    (total_values * 2) as u32,
-                );
-                let prev_len = packed_data.len();
-                packed_data.resize(prev_len + total_values * 2, 0);
-                u8_view.copy_to(&mut packed_data[prev_len..]);
-            }
-
-            sweep_meta_vec.push(VolumeRenderSweepMeta {
-                elevation_deg: header.mean_elevation as f64,
-                azimuth_count: header.azimuth_count,
-                gate_count: header.gate_count,
-                first_gate_km: header.first_gate_range_km,
-                gate_interval_km: header.gate_interval_km,
-                max_range_km: header.max_range_km,
-                data_offset,
-                scale: header.scale as f64,
-                offset: header.offset as f64,
-            });
-
-            data_offset += total_values as u32;
-        }
-
-        let total_ms = t_total.elapsed().as_secs_f64() * 1000.0;
-
-        log::info!(
-            "render_volume: {} sweeps, {} values packed ({:.1}KB, u{}) in {:.1}ms",
-            sweep_meta_vec.len(),
-            data_offset,
-            packed_data.len() as f64 / 1024.0,
-            word_size * 8,
-            total_ms,
-        );
-
-        // Serialize scalar/struct fields, then attach the packed buffer separately
-        let response = VolumeRenderResponse {
-            sweep_count: sweep_meta_vec.len() as u32,
-            word_size,
-            sweep_meta: sweep_meta_vec,
-            product: product_str,
-            total_ms,
-        };
-        let result = serde_wasm_bindgen::to_value(&response)
-            .map_err(|e| JsValue::from_str(&format!("Failed to serialize response: {}", e)))?;
-
-        // ArrayBuffer must be set directly for zero-copy transfer
-        let packed_u8 = js_sys::Uint8Array::from(&packed_data[..]);
-        let packed_buffer = packed_u8.buffer();
-        js_sys::Reflect::set(&result, &"buffer".into(), &packed_buffer).ok();
-
-        Ok(result)
-    })
-}
-
 // ---------------------------------------------------------------------------
 // Per-chunk incremental ingest
 // ---------------------------------------------------------------------------
@@ -856,23 +184,23 @@ pub fn worker_render_volume(params: wasm_bindgen::JsValue) -> js_sys::Promise {
 /// Accumulator for per-chunk ingest. Holds decoded radials across chunks
 /// until an elevation is complete, then flushes sweep blobs to IDB.
 #[allow(dead_code)]
-struct ChunkAccumulator {
-    scan_key: ScanKey,
-    site_id: String,
-    all_radials: Vec<::nexrad::model::data::Radial>,
-    radial_metas: Vec<(i64, u8, f32, f32)>,
-    completed_elevations: std::collections::HashSet<u8>,
-    last_elevation_number: Option<u8>,
-    vcp: Option<ExtractedVcp>,
-    has_vcp: bool,
-    total_chunks: u32,
-    total_size_bytes: u64,
-    file_name: String,
-    timestamp_secs: i64,
+pub(super) struct ChunkAccumulator {
+    pub scan_key: ScanKey,
+    pub site_id: String,
+    pub all_radials: Vec<::nexrad::model::data::Radial>,
+    pub radial_metas: Vec<(i64, u8, f32, f32)>,
+    pub completed_elevations: std::collections::HashSet<u8>,
+    pub last_elevation_number: Option<u8>,
+    pub vcp: Option<ExtractedVcp>,
+    pub has_vcp: bool,
+    pub total_chunks: u32,
+    pub total_size_bytes: u64,
+    pub file_name: String,
+    pub timestamp_secs: i64,
 }
 
 thread_local! {
-    static CHUNK_ACCUM: std::cell::RefCell<Option<ChunkAccumulator>> =
+    pub(super) static CHUNK_ACCUM: std::cell::RefCell<Option<ChunkAccumulator>> =
         const { std::cell::RefCell::new(None) };
 }
